Detail texture geometry fixes, and offscreen loading tweak
Changes in support of Simon Wilson's work on improving the
Books carousel appearance. Provide an option to center the
detail texture with respect to the card, as opposed to the
default left-alignment. Draw the detail texture based on the
bounding box of the card, not just on the left coordinate.
Add an option to specify how many offscreen cards are
maintained in memory (as opposed to just one on each side).

Change-Id: I696f7315423440a9cecb503f07be96ffee96424f