Fix 3099806: Prevent glitch when scrolling to end of Carousel.
This fixes a bug where some activities were seeing a
"bounce" at the end of the Carousel list. The issue would
happen because the velocity wasn't reset before the
overscroll animation completed, which meant any call causing
the screen to refresh would run doPhysics() which would cause
the overscroll animation to run.

Change-Id: I2e19603cca3e0ca3374010a762429b8d79804b9e