Carousel performance improvements.
This change disables depth and changes the background bitmap
for the test case to 256x256 instead of 512x512. With these
two changes, Carousel consistently achieves 87fps with no
background and 78fps with. Both tests were with sync-to-vblank
disabled.

It also cleans up the program store names to have more
meaningful names as we add new states.

Upload after rebase.

Change-Id: I6008fbd89fbe924dbea10792b3f787fe69586fa0