Tap to stop the carousel
When the carousel is moving faster than 15 degrees per
second, a tap will stop the motion but not be interpreted
as a selection. Slower than that, it will be considered a
selection.

This is an improvement on an earlier implementation because
it lets through selections when the carousel is still
moving slowly.

Bug: 3179984
Change-Id: I27b55cf88e20d4ef1383ec0e4144f65a101f37fc