Arrange cards based on default matrix
Previously we were arranging cards on the screen using the
individual card's matrix. When the card matrices are
different, however, this leads to an inconsistent layout.
Instead, use the default card matrix for layout, while
using the individaul card matrix (which, when not set
separately, is the same as the default card matrix) for
handling drawing the individual card.

Bug: 3365044
Change-Id: Ibb7ad7760131d2c649cc0882b784c9d2e73a737f