Fix filtering bug and reduce area that causes delete of chip.
The bug was: when the user backspaced from the end of the multiautocompletefield
to a chip, the view would bring up a popup that filtered against hte contents
of the chip. This fixes that by specifically not filtering on chip contents.

Also, dont allow taps past the end of a chip to delete the chip.
Change-Id: Idbb41756c9efd544846802ddad8c67545b8ade1f