Create chips when text is appended.
When text is appended to the multiautocompletetextfield,
we need to make sure we make chips of the added text.
Since in many cases it will be added before the view
is measured, we may need to wait for onSizeChanged to get
called. This stores chip info for later use in creating chips
once we can measure the width / height of the view.
Change-Id: Ifc9b8bdf089fb0293bb76ea9cd7aeacfc3f9da87