Mke sure we commit chips that are being edited
The isse was that text editing of a chip now holds the chip
in the selected chip field. This makes sure that we commit any chip
that has an invalid id before continuing.

Fixes bug:5277118 Chip layout is weird if there are "+x more" text and a recipient not in your contact list

Change-Id: I7480a3677e1fa1d17e365c3057dd74ae46ede858