Enabled optional support for showing progress.
Added some progress bars to the "empty" view
so that implementors of the API can show the
progress of things like downloads or loading.

Additionally, added a retry button and an optional
text that allows you to indicate status to the user.

Due to a bug in the framework, we actually use two
progress bars (one determinate and one indeterminate)
to update our status. Created a wrapper that allows
the controlling of both progress bars in sync.

Change-Id: I710e06317948d376d445abf1603545982ac8479d