Scope changes and removal of unnessecary variable
Adds isFullScreen() and setPhotoIndex() protected methods to
PhotoViewActivity in order to allow subclasses access to this
information.

Removes mDataValid from BaseCursorPagerAdapter since it was hiding how
simple the checks actually were (it was just mCursor != null)

A nullcheck has also been added to BaseFragmentPagerAdapter in
instantiateItem. in the long term, BaseCursorPagerAdapter should not be
so fragile, and should not surface nulls, but that change could be quite
involved, so should probably not be rolled into this change.

Change-Id: I4889c5ae2a764660926c70da25b85565f4600188