Do not show loading bar after thumbnail has loaded
After talking to the designers for Google Now, who have said they talked
to the designers from gallery app, it appears that we should not show
the loading spinner once the thumbnail has loaded.

Let me know if this is orthogonal to the behaviour you would like in the
gmail app, and I can write a more involved patch with an overridable
callback to decide wether to show the spinner or not.

Change-Id: I77a3f2e2487d469ab47e45c7a94942f831528d41