Ensure we chipify the original text
We were tokenizing the addresses, then attempting to chipify what we
believed was the ideal way to represent that address as a string. If
this did not match the actual text, it resulted in not everything
being turned into a chip.

By actually using the original text to determine what we make a chip,
we can ensure that everything is always turned into a chip.

Bug: 8321211
Change-Id: I5e7b4d57b6b8895ae5211818451075f461c14eaf