Consolidate parameter caching in camera1 portability agent
The existing implementation of the AndroidCameraAgentImpl uses
a 'mParamsToSet' upon which is projects requested settings
changes in applySettings(). But, this mParamsToSet is initialized
only upon camera open! So 'hidden' parameters like video-size are
stuck in the value upon open. Instead, they should be current to
the last call to refreshSettings. To simplify, removed this and
another copy of parameters (mParameters), introduced one cache
class, and use it throughout the agent. Tested and app seems to
work OK in general with the change, and specifically confirmed that
video size parameters are not reverted after start, as was
observed before the change.

Bug: 16991558
Change-Id: I8f2749b8ab2ccc2cb858b8cd844330c5a482f79c