camera2-portability: Fix zoom occasionally not working
Cache the last valid CameraSettings object

The camera2 implementation of CameraProxy#getSettings() now returns the most
recent valid CameraSettings object that was provided to it (or the template if
no valid cached instance is available). This more closely mimics the camera1
implementation's behavior, and resolves issues where the app relies on finding
recent settings there in corner cases.

Bug: 17016658
Change-Id: I057feec3f857428dd210300a05f22cafa9f5ba34