Remove apache dependency from android-common.
This static library was relying on HttpDateTime via AndroidHttpClient.
(and is the only user of this API in our tree). Regrettably, the most
reasonable way forward seems to be copy that class into frameworks/ex.

Using reflection is possible too, but it's riskier since we have to
ensure this class exists unmodified all the way back to froyo.

See API council request b/2553589 for why this was in AndroidHttpClient
in the first place.

Also, this change sets LOCAL_SDK_VERSION := 8 for tests to match the
library version.

bug: 18027885

Change-Id: I9c473f60f0314758120e0d6c6f9b56e4d77ba658